To Do List App 0.1

+++ app/todoapp/views.py
@@ -6,16 +6,22 @@ from .models import TodoListItem
 
 def todoappView(request):
     all_todo_items = TodoListItem.objects.all()
-    return render(request, 'todolist.html',
+    return render(request, 'todoapp/todolist.html',
     {'all_items':all_todo_items})
 
 def addTodoView(request):
-    x = request.POST['content']
-    new_item = TodoListItem(content = x)
-    new_item.save()
+    x = str(request.POST['content']).strip()
+    if x != "":
+        new_item = TodoListItem(content = x)
+        new_item.save()
     return HttpResponseRedirect('/todoapp/')
 
-def deleteTodoView(request, i):
+def actionTodoView(request, te, i):
     y = TodoListItem.objects.get(id= i)
-    y.delete()
+
+    if te == 'status':
+        y.status = not y.status
+        y.save_base()
+    elif te == 'delete':
+        y.delete()
     return HttpResponseRedirect('/todoapp/')
